TSAPPS-231 Use recipient name in offers import
It will be more comfortable for customer to set up recipient's name instead of ID

In real case it should not be more than 2-3 recipients per supplier, so pairs name/id can be added into config.
As fast solution

Tech debt: raw config file is too big now, it was splitted
