--- config/rawRules.go
@@ -1,9 +1,0 @@
-package config
-
-type RawRules struct {
-	Map map[string]string `yaml:"column-mappings" validate:"required"`
-}
-
-func (c *RawRules) ToConfig() map[string]string {
-	return c.Map
-}
